feat(events): Shows extra info on all the tabs (#3012)
