+++ components/Events/EventsRequestEvent.vue
@@ -1,0 +1,37 @@
+<template>
+  <div class="request-event__section">
+    <h3>Start an event</h3>
+    <p class="request-event__section__description">
+      We can help you bring Qiskit experts to your campus for guest lectures,
+      hackathons, and other events.
+    </p>
+    <UiCta v-bind="eventRequestLink" />
+  </div>
+</template>
+
+<script setup lang="ts">
+import { EVENT_REQUEST_LINK } from "~/constants/appLinks";
+
+const eventRequestLink = EVENT_REQUEST_LINK;
+</script>
+
+<style lang="scss" scoped>
+@use "~/assets/scss/carbon.scss";
+
+.request-event {
+  &__section {
+    margin-top: carbon.$spacing-10;
+    margin-bottom: carbon.$spacing-10;
+
+    &__description {
+      margin-top: carbon.$spacing-06;
+      margin-bottom: carbon.$spacing-07;
+      max-width: 20rem;
+
+      @include carbon.breakpoint-up(lg) {
+        max-width: 24rem;
+      }
+    }
+  }
+}
+</style>
