+++ pages/events/index.vue
@@ -1,5 +1,5 @@
 <template>
-  <div>
+  <div class="event-page">
     <UiPageHeaderFixed>
       Join
       <UiTypewriterEffect
@@ -20,14 +20,15 @@
           </bx-tabs>
         </client-only>
       </div>
-      <div v-if="showCalendar" class="event-page__calendar">
+      <div v-if="showCalendar">
         <iframe
-          class="airtable-embed"
+          class="event-page__calendar"
           src="https://airtable.com/embed/shrzmTpiOo1Ye8Nrs?backgroundColor=purple&viewControls=on"
           width="100%"
           height="560"
-          allowtransparency="true"
         />
+        <EventsFollowCalendar />
+        <EventsRequestEvent />
       </div>
       <UiFiltersResultsLayout v-else>
         <template #filters-on-m-l-screen>
@@ -92,62 +93,8 @@
           </div>
         </template>
         <template #extra-info>
-          <div class="event-page__section">
-            <h3>Follow our event calendar</h3>
-            <p class="event-page__section__description">
-              Stay up to date with all of our scheduled events by following our
-              calendar. You can view the calendar by visiting
-              <UiLink v-bind="qiskitCalendarLink">{{
-                qiskitCalendarLink.label
-              }}</UiLink
-              >, or subscribe to it by adding to the calendar app of your
-              choice.
-            </p>
-            <div class="event-page__tabs">
-              <bx-tabs value="Google">
-                <bx-tab
-                  v-for="{ name } in calendarsInstructions"
-                  :id="`tab-${name}`"
-                  :key="name"
-                  :target="`panel-${name}`"
-                  :value="name"
-                  class="event-page__tab"
-                >
-                  {{ name }}
-                </bx-tab>
-              </bx-tabs>
-            </div>
-            <div
-              v-for="{ name, instructions } in calendarsInstructions"
-              :id="`panel-${name}`"
-              :key="name"
-              role="tabpanel"
-              :aria-labelledby="`tab-${name}`"
-              hidden
-            >
-              <p class="event-page__sync">
-                Start by copying the calendar subscription link
-                <bx-code-snippet
-                  copy-button-feedback-text="Copy calendar sync url"
-                  type="inline"
-                >
-                  <span>{{ qiskitCalendarSyncLink }}</span>
-                </bx-code-snippet>
-              </p>
-              <component :is="instructions" class="event-page__instructions" />
-            </div>
-          </div>
-          <div class="event-page__section">
-            <h3>Start an event</h3>
-            <p class="event-page__section__description">
-              We can help you bring Qiskit experts to your campus for guest
-              lectures, hackathons, and other events.
-            </p>
-            <UiCta
-              :label="eventRequestLink.label"
-              :url="eventRequestLink.url"
-            />
-          </div>
+          <EventsFollowCalendar />
+          <EventsRequestEvent />
         </template>
       </UiFiltersResultsLayout>
     </div>
@@ -163,10 +110,7 @@ import {
   WORLD_REGION_OPTIONS,
   COMMUNITY_EVENT_TYPE_OPTIONS,
 } from "~/types/events";
-import EventsCalendarsAppleInstructions from "~/components/Events/Calendars/EventsCalendarsAppleInstructions.vue";
-import EventsCalendarsGoogleInstructions from "~/components/Events/Calendars/EventsCalendarsGoogleInstructions.vue";
-import EventsCalendarsOutlookInstructions from "~/components/Events/Calendars/EventsCalendarsOutlookInstructions.vue";
-import { EVENT_REQUEST_LINK, GeneralLink } from "~/constants/appLinks";
+
 import rawPastEvents from "~/content/events/past-community-events.json";
 import rawUpcomingEvents from "~/content/events/upcoming-community-events.json";
 
@@ -182,34 +126,6 @@ const upcomingEvents = rawUpcomingEvents as CommunityEvent[];
 useHead({
   title: "Qiskit Events",
 });
-
-const calendarsInstructions = [
-  {
-    name: "Google",
-    instructions: EventsCalendarsGoogleInstructions,
-  },
-  {
-    name: "Outlook",
-    instructions: EventsCalendarsOutlookInstructions,
-  },
-  {
-    name: "Apple",
-    instructions: EventsCalendarsAppleInstructions,
-  },
-];
-
-const qiskitCalendarSyncLink = "https://qisk.it/calendar-sync";
-
-const qiskitCalendarLink: GeneralLink = {
-  url: "https://qisk.it/calendar",
-  label: "https://qisk.it/calendar",
-  segment: {
-    cta: "qiskit-calendar",
-    location: "events-page",
-  },
-};
-
-const eventRequestLink = EVENT_REQUEST_LINK;
 
 const emptyCard = {
   title: "No events found",
@@ -426,26 +342,31 @@ const selectTab = (selectedTab: string) => {
     }
   }
 
-  &__section {
+  &:deep(.event-page__section) {
     margin-top: carbon.$spacing-10;
     margin-bottom: carbon.$spacing-10;
+  }
 
-    &__description {
-      margin-top: carbon.$spacing-06;
-      margin-bottom: carbon.$spacing-07;
-      max-width: 20rem;
+  &:deep(.event-page__section__description) {
+    margin-top: carbon.$spacing-06;
+    margin-bottom: carbon.$spacing-07;
+    max-width: 20rem;
 
-      @include carbon.breakpoint-up(lg) {
-        max-width: 24rem;
-      }
+    @include carbon.breakpoint-up(lg) {
+      max-width: 24rem;
     }
   }
 
-  &__tab {
+  &:deep(.event-page__tabs) {
+    margin-top: carbon.$spacing-07;
+    margin-bottom: carbon.$spacing-09;
+  }
+
+  &:deep(.event-page__tab) {
     padding-top: carbon.$spacing-06;
   }
 
-  &__instructions {
+  &:deep(.event-page__instructions) {
     padding-left: carbon.$spacing-06;
   }
 }
